backend/apps/integrations/services/stt_service.py
from google.cloud import speech
from django.conf import settings
import os

class STTService:
    """
    Speech-to-Text service using Google Cloud Speech API
    TODO: Add Google Cloud credentials file path in .env
    """
    
    def __init__(self):
        # TODO: Set up Google Cloud credentials
        # os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = settings.GOOGLE_CLOUD_CREDENTIALS
        self.client = None
        # self.client = speech.SpeechClient()
    
    def transcribe_audio(self, audio_content, language_code='en-US'):
        """
        Transcribe audio content to text
        
        Args:
            audio_content: Audio file bytes
            language_code: Language code (default: en-US)
            
        Returns:
            dict: {transcript: str, confidence: float}
        """
        # TODO: Uncomment when Google Cloud credentials are configured
        # try:
        #     audio = speech.RecognitionAudio(content=audio_content)
        #     config = speech.RecognitionConfig(
        #         encoding=speech.RecognitionConfig.AudioEncoding.LINEAR16,
        #         sample_rate_hertz=16000,
        #         language_code=language_code,
        #         enable_automatic_punctuation=True,
        #     )
        #     
        #     response = self.client.recognize(config=config, audio=audio)
        #     
        #     if response.results:
        #         result = response.results[0]
        #         alternative = result.alternatives[0]
        #         return {
        #             'transcript': alternative.transcript,
        #             'confidence': alternative.confidence
        #         }
        #     
        #     return {'transcript': '', 'confidence': 0.0}
        # 
        # except Exception as e:
        #     raise Exception(f"STT failed: {str(e)}")
        
        # Placeholder response
        return {
            'transcript': 'This is a placeholder transcription. Configure Google Cloud Speech API.',
            'confidence': 0.95
        }
    
    def transcribe_stream(self, audio_generator, language_code='en-US'):
        """
        Transcribe audio stream in real-time
        
        Args:
            audio_generator: Generator yielding audio chunks
            language_code: Language code
            
        Yields:
            dict: {transcript: str, is_final: bool, confidence: float}
        """
        # TODO: Implement streaming STT
        # config = speech.RecognitionConfig(...)
        # streaming_config = speech.StreamingRecognitionConfig(config=config, interim_results=True)
        # requests = (speech.StreamingRecognizeRequest(audio_content=chunk) for chunk in audio_generator)
        # responses = self.client.streaming_recognize(streaming_config, requests)
        # 
        # for response in responses:
        #     for result in response.results:
        #         yield {
        #             'transcript': result.alternatives[0].transcript,
        #             'is_final': result.is_final,
        #             'confidence': result.alternatives[0].confidence
        #         }
        
        yield {
            'transcript': 'Streaming placeholder',
            'is_final': True,
            'confidence': 0.95
        }
