backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# MongoDB
pymongo==4.6.0
mongoengine==0.27.0  # Better MongoDB ODM for Django

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# AI & ML
openai==1.3.0
# TODO: Add your OpenAI API key in .env file

# CV Parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Speech Processing
google-cloud-speech==2.21.0
# TODO: Add Google Cloud credentials file path in .env
azure-cognitiveservices-speech==1.31.0
# TODO: Add Azure Speech API key in .env

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.14

# Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Video Processing
ffmpeg-python==0.2.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pillow==10.1.0
requests==2.31.0

# Development
python-dotenv==1.0.0
